requirements.txt
fastapi
uvicorn
streamlit
pandas
groq
requests
python-dotenv
